add scripts to get memory usage
But each of these scripts has problems... the one that uses smap
often gets a permission denied error. The psutil script does not
seem to measure the total memory. And if we measure child memory
as well, then we seem to overestimate memory, probably because some
memory pages are shared yet we don't account for this.